+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The `pylexibank` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## unreleased
+
+- Removed lexibank db functionality.
+
+
 ## 3.5.0 - 2024-07-04
 
 - Changed behaviour of makecldf so it does not remove languages or parameters explicitly added by the user.